requirements.txt
-i https://pypi.org/simple
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
boto3==1.16.4
botocore==1.19.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4 ; sys_platform == 'win32'
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum-compat==0.0.3
gitdb==4.0.5
gitpython==3.1.11
idna==2.10
importlib-metadata==2.0.0 ; python_version < '3.8'
ipykernel==5.3.4 ; python_version >= '3.4'
ipython-genutils==0.2.0
ipython==7.18.1 ; python_version >= '3.3'
ipywidgets==7.5.1
jedi==0.17.2
jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
markupsafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyarrow==2.0.0
pycparser==2.20
pydeck==0.4.1
pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==228 ; sys_platform == 'win32'
pywinpty==0.5.7 ; os_name == 'nt'
pyzmq==19.0.2
requests==2.24.0
s3transfer==0.3.3
send2trash==1.5.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1 ; python_version >= '3.0'
streamlit==0.69.2
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
traitlets==5.0.5
typing==3.7.4.3
tzlocal==2.1
urllib3==1.25.11 ; python_version != '3.4'
validators==0.18.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.3.2
